fixed bugs in ap and liquidation - decimal values are not being include in the total computation

+++ resources/views/livewire/create-account-payable-liquidation-step1-component.blade.php
@@ -212,7 +212,7 @@
                                             @enderror
                                         </x-td>
                                         <x-td>
-                                           {{ number_format((int)$particular->quantity * (int)$particular->price, 2) }}
+                                           {{ number_format((double)$particular->quantity * (double)$particular->price, 2) }}
                                         </x-td>
                                       
                                         <x-td>
